refactor(map): delete dead osm_retry_client.dart + fold Retry-After into RetryingTileHttpClient
Closes #1654. osm_retry_client.dart (OsmRetryClient, 127 lines) was a dead
near-duplicate of RetryingTileHttpClient — same 429/5xx/SocketException
retry policy, same jittered exponential backoff — with zero production
call sites. Two retry clients, one in use.

Deleted osm_retry_client.dart and its test. Its one capability the
surviving client lacked — honouring a `Retry-After` header (OSM's
preferred rate-limit signal on a 429) — is folded into
RetryingTileHttpClient: `_nextDelay` now takes the failing response
headers, and a `Retry-After` value (delay-seconds or HTTP-date) wins over
exponential backoff when present.
